# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MS App",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/task": {
      "get": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/task/update": {
      "put": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task/delete": {
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task/comment/add": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task/comment/update": {
      "put": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "commentId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task/comment/delete": {
      "delete": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "commentId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task/history/add": {
      "post": {
        "tags": [
          "Task"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/timelog/upload": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to upload timelog.",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/timelog/all": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to get all user timelog.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/timelog": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to get the specific user timelog.",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/timelog/update": {
      "put": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to update specific user timelog.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/timelog/delete": {
      "delete": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to delete specific user timelog.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/timelog/totalhours": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to get the specific user timelog total hours.",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/timelog/timeloghours": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to get the specific user timelog hours.",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/timelog/projecthours": {
      "get": {
        "tags": [
          "Timelog"
        ],
        "description": "Endpoint to get the specific user project timelog.",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/todo": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/todo/update": {
      "put": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/todo/delete": {
      "delete": {
        "tags": [
          "Todo"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}